Refactor sts func into separate file
