+++ coredb-operator/src/controller.rs
@@ -1,18 +1,10 @@
 use crate::{telemetry, Error, Metrics, Result};
 use chrono::{DateTime, Utc};
 use futures::{future::BoxFuture, FutureExt, StreamExt};
-use k8s_openapi::{
-    api::{
-        apps::v1::{StatefulSet, StatefulSetSpec},
-        core::v1::{
-            Container, ContainerPort, EnvVar, PersistentVolumeClaim, PersistentVolumeClaimSpec, PodSpec,
-            PodTemplateSpec, ResourceRequirements,
-        },
-    },
-    apimachinery::pkg::{api::resource::Quantity, apis::meta::v1::LabelSelector},
-};
+
+use crate::statefulset::create_sts;
 use kube::{
-    api::{Api, ListParams, ObjectMeta, Patch, PatchParams, ResourceExt},
+    api::{Api, ListParams, Patch, PatchParams, ResourceExt},
     client::Client,
     runtime::{
         controller::{Action, Controller},
@@ -24,7 +16,7 @@ use kube::{
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use serde_json::json;
-use std::{collections::BTreeMap, sync::Arc};
+use std::sync::Arc;
 use tokio::{sync::RwLock, time::Duration};
 use tracing::*;
 
@@ -106,85 +98,9 @@ impl CoreDB {
             .await
             .map_err(Error::KubeError)?;
         // create statefulset
-        self.create_sts(ctx).await.expect("error creating statefulset");
+        create_sts(self, ctx).await.expect("error creating statefulset");
         // If no events were received, check back every minute
         Ok(Action::requeue(Duration::from_secs(60)))
-    }
-
-    async fn create_sts(&self, ctx: Arc<Context>) -> Result<(), Error> {
-        let client = ctx.client.clone();
-        let ns = self.namespace().unwrap();
-        let name = self.name_any();
-        let mut labels: BTreeMap<String, String> = BTreeMap::new();
-        let mut pvc_requests: BTreeMap<String, Quantity> = BTreeMap::new();
-        let sts_api: Api<StatefulSet> = Api::namespaced(client, &ns);
-        let oref = self.controller_owner_ref(&()).unwrap();
-        labels.insert("app".to_string(), "coredb".to_string());
-        pvc_requests.insert("storage".to_string(), Quantity("8Gi".to_string()));
-
-        let sts: StatefulSet = StatefulSet {
-            metadata: ObjectMeta {
-                name: Some(name.to_owned()),
-                namespace: Some(ns.to_owned()),
-                labels: Some(labels.clone()),
-                owner_references: Some(vec![oref]),
-                ..ObjectMeta::default()
-            },
-            spec: Some(StatefulSetSpec {
-                replicas: Some(self.spec.replicas),
-                selector: LabelSelector {
-                    match_expressions: None,
-                    match_labels: Some(labels.clone()),
-                },
-                template: PodTemplateSpec {
-                    spec: Some(PodSpec {
-                        containers: vec![Container {
-                            env: Option::from(vec![EnvVar {
-                                name: "POSTGRES_PASSWORD".parse().unwrap(),
-                                value: Some("password".parse().unwrap()),
-                                value_from: None,
-                            }]),
-                            name: name.to_owned(),
-                            image: Some("docker.io/postgres:15".to_owned()),
-                            ports: Some(vec![ContainerPort {
-                                container_port: 5432,
-                                ..ContainerPort::default()
-                            }]),
-                            ..Container::default()
-                        }],
-                        ..PodSpec::default()
-                    }),
-                    metadata: Some(ObjectMeta {
-                        labels: Some(labels),
-                        ..ObjectMeta::default()
-                    }),
-                },
-                volume_claim_templates: Option::from(vec![PersistentVolumeClaim {
-                    metadata: ObjectMeta {
-                        name: Some("data".to_string()),
-                        ..ObjectMeta::default()
-                    },
-                    spec: Some(PersistentVolumeClaimSpec {
-                        access_modes: Some(vec!["ReadWriteOnce".to_owned()]),
-                        resources: Some(ResourceRequirements {
-                            limits: None,
-                            requests: Some(pvc_requests),
-                        }),
-                        ..PersistentVolumeClaimSpec::default()
-                    }),
-                    status: None,
-                }]),
-                ..StatefulSetSpec::default()
-            }),
-            ..StatefulSet::default()
-        };
-
-        let ps = PatchParams::apply("cntrlr").force();
-        let _o = sts_api
-            .patch(&name, &ps, &Patch::Apply(&sts))
-            .await
-            .map_err(Error::KubeError)?;
-        Ok(())
     }
 
     // Finalizer cleanup (the object was deleted, ensure nothing is orphaned)
