+++ coredb-operator/src/lib.rs
@@ -33,3 +33,4 @@ mod metrics;
 pub use metrics::Metrics;
 
 #[cfg(test)] pub mod fixtures;
+mod statefulset;
